Fixed an issue where notifications were invisible on AOD
Because of the new screen off animation, the shelf would
be positioned wrongly and therefore the icons would
be clipped.
We'll have to revisit this positioning, when we do the real AOD
layout, for now we place the invisible shelf at the bottom
of the last notification.

Change-Id: If2bd7c71957585ac73ace004221e48c5b5f134af
Fixes: 76160478
Test: add ambient notification,  observe  normal clipping
(cherry picked from commit cb8e240e95d6f5135b0eaa7de3190a649cf2a7cf)