Show work notifications in the shade unless on keyguard.
1. Make work profile notifications redacted but visible
if only work profile is locked (i.e. not on the keyguard).
Previously if the user disables notifications on the lockscreen,
work notifications will disappear from notification shade when
the profile is locked even if the device itself is not locked.
To achieve this NotificationLockscreenUserManager now checks
notification status for a given user rather than system-wide.
When the keyguard is locked, system-wide setting will still prevent
work notification from being visible.

2. Fix bug introduced by previous CL, since isCurrentProfile
is alwoays true for mCurrentUser, it caused notifications
to be visible on keyguard when they shouldn't.

Test: manual, tried combinations of device and profile settings
Test: runtest systemui
Bug: 64829587
Change-Id: I104e938b15897a909c086c62e273f922b96ba7db