Fix CaptivePortalLoginActivity null String crash
In getHeaderTitle if NetworkInfo.getExtraInfo() returns null the call to
replaceAll will result in an NPE and if the string weren't null an empty
string would set the title to "Sign in to ".

Add a TextUtils.isEmpty check before displaying the formatted title to
temporarily fix the NPE until an alternative method to obtain SSID is
implemented.

Bug: 77114259
Test: Manually launching with a captive portal network.

Change-Id: I3ce9a6dc65cceb9344700803966d4f0aa12d894c