Opalayout can check quickstep icons without needing navbarview (1/2)
Opalayout should not need NavigationBarView to determine if the icon
should be for quickstep icons. It does not need to know if recents is
disabled because of lock tasks and so the check is now in
OverviewProxyService. Then when Opalayout needs to layout the home
button it should be able to get the information regardless if
NavigationBarView is available or not.

Bug: 76210393
Fixes: 76434966
Fixes: 74071465
Test: change language in settings

Change-Id: I246175c3e2d5c990da6823df4108d48522b258bd