Calculate multipath quota excluding current day.
The current calculation includes data usage for the current day,
which causes the returned daily quota to change during the day.
This makes it difficult to use for callers that need a fixed
data limit for the day.

Current usage in MultipathPolicyTracker (the only caller) is
incorrect because of this, as it assumes that quota is not constantly
changing.

Test: Tests in go/ag/3803612 pass
Bug: b/72631572
Bug: b/72877610
Change-Id: I9b3edde2dee7c7479d428e2bf2b73297afed3e6b