Move PrintRecommendationService to systemAPI
It is using system APIs, but we might want to update it via the play
store.

Hence, replace all internal APIs used with support lib APIs. There is
no replacement for @Immutable, but this is just a informational
annotation, so it can just be removed.

As system-api is tied to a specific platform version, remove all
minSDK/targetSDK settings in the manifest.

Change-Id: Ia8f9c718fd688db295c7bc3bfe05e2074f9bd413
Fixes: 78162439
Test: Looked at print service recommendations