Thaw input dispatching immediately after unfreezing.
Currently, we thaw after taking a number of steps post-unfreezing,
including setting up screen animations for rotation. If there is an
exception in this intermediate code, the input dispatching will not
be thawed even though the display has unfrozen.

This changelist moves thawing to immediately after unfreezing to
prevent it from being skipped due to an exception.

Bug: 78301580
Test: atest CtsActivityManagerDeviceTestCases
Change-Id: I32a3ca909fe1a495b361f9e5ba7b3a158837bc3a