Fixed issue where color attributes used for window background Activity themes would not be applied
Updated windowBackground attribute in attrs.xml to take either a
reference or a color parameter. Updated PhoneWindow to consume the
drawable defined from the TypedArray. Added test to PhoneWindowTest to
handle verifying if a theme has a background color defined that it is
applied properly.

Fixes: 70268864
Test: Re-ran core tests on PhoneWindowTest with an updated test to
verify color literals are applied properly

Change-Id: I77083d5ee366ca0236633ac7039d2a5acf127fe7