Fix night light and COLOR_MODE_AUTOMATIC
The native mode methods in DisplayTransformManager were used to
decide whether the color matrix works in linear or gamma space.  For
COLOR_MODE_AUTOMATIC, the color matrix works in lienar instead of
gamma space.

Fix the native mode methods to be more clear about the intention to
avoid future breakage.

Bug: 79257138
Test: manual and atest FrameworksServicesTests:ColorDisplayServiceTest
Change-Id: Ied60a62bd7330e87c3207f2183d7e83af9e51fb7